hr-breaker-main/src/hr_breaker/services/pdf_parser.py
"""PDF text extraction using PyMuPDF."""

from pathlib import Path

import fitz  # pymupdf


def extract_text_from_pdf(pdf_path: Path) -> str:
    """Extract text from PDF file.

    Args:
        pdf_path: Path to PDF file

    Returns:
        Extracted text content
    """
    doc = fitz.open(pdf_path)
    text_parts = []

    for page in doc:
        text_parts.append(page.get_text())

    doc.close()
    return "\n".join(text_parts)
